requirements.txt
fastapi
uvicorn
openai
pdfminer.six
python-multipart
